requirements.txt
streamlit
pydicom
Pillow
pandas
pylibjpeg 
pylibjpeg-libjpeg
python-gdcm